Version2 of Macro 1 (Use this one)
Most updated version (15th January 2024).

The previous version had some bugs and it didn't work unless the dye of the fluorescence channel was red.
Now it works with blue and green too.

I have added the option to select if the data is a multichannel image or not, because in some cases only a fluorescence channel was capture and the researcher wanted to analyse these images so now there is an additional question in the dialog box asking if the data is a multichannel image or not. 

The appropriate comments have been added.